pyproject.toml
[project]
name = "california-housing-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "mlflow>=3.8.1",
    "numpy>=2.4.0",
    "optuna>=4.6.0",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "types-pyyaml>=6.0.12.20250915",
    "xgboost>=3.1.2",
]
